# doc/openapi3.yaml
openapi: 3.0.4
info:
  title: Orticello API
  version: 1.0.0
  description: API for managing community gardens (orti urbani) in Trento
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://orticelloapi.onrender.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Bando:
      type: object
      required:
        - titolo
        - data_inizio
        - data_fine
        - messaggio
      properties:
        _id:
          type: string
          description: MongoDB ObjectId (generated automatically)
          readOnly: true
          example: 507f1f77bcf86cd799439020
        titolo:
          type: string
          description: Competition announcement title
          example: Bando per assegnazione lotti 2026
        data_inizio:
          type: string
          format: date
          description: Application period start date
          example: 2026-01-01
        data_fine:
          type: string
          format: date
          description: Application period end date
          example: 2026-01-31
        messaggio:
          type: string
          description: Competition announcement message
          example: Sono aperti i bandi per l'assegnazione dei lotti per l'anno 2026
        link:
          type: string
          description: Optional link to full announcement
          example: https://www.comune.trento.it/bandi/orti-2026
...
paths:
...
  /api/v1/bandi:
    get:
      summary: Get list of announcements
      description: Retrieve a list of all announcements. Use query param 'active=true'
        to filter only current ones.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      parameters:
        - in: query
          name: active
          schema:
            type: boolean
          description: Se impostato a true, restituisce solo i bandi non scaduti (ordinati
            per scadenza).
      responses:
        "200":
          description: Successfully retrieved list of bandi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bando"
        "500":
          description: Error retrieving bandi
    post:
      summary: Create a new announcement
      description: Create a new bando. Only accessible by Comune.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Bando"
      responses:
        "201":
          description: Bando created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bando"
        "400":
          description: Invalid input data
        "500":
          description: Error creating bando
  /api/v1/bandi/attivi:
    get:
      summary: Get active announcements
      description: Retrieve only bandi that have already started and not yet ended
        (data_inizio <= now <= data_fine). Results are sorted by end date.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      responses:
        "200":
          description: Successfully retrieved list of active bandi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bando"
        "500":
          description: Error retrieving active bandi
  /api/v1/bandi/{id}:
    get:
      summary: Get announcement by ID
      description: Retrieve a single bando by ID.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved bando
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bando"
        "404":
          description: Bando not found
        "500":
          description: Error retrieving bando
    put:
      summary: Update announcement by ID
      description: Update an existing bando. Only accessible by Comune.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Bando"
      responses:
        "200":
          description: Bando updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bando"
        "404":
          description: Bando not found
        "500":
          description: Error updating bando
    delete:
      summary: Delete announcement by ID
      description: Remove an existing bando. Only accessible by Comune.
      security:
        - bearerAuth: []
      tags:
        - Bandi
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bando deleted successfully
        "404":
          description: Bando not found
        "500":
          description: Error deleting bando
...